feat(87): detail 디렉토리 정리
