+++ src/components/detail/EventNearHere.tsx
@@ -1,13 +1,13 @@
 import React from "react";
 import { useQuery } from "react-query";
-import { useParams } from "react-router-dom";
-import { StyledEventNearHere } from "../../styles/eventNearHereStyle";
-import EventNearHereList from "./EventNearHereLIst";
+import { useNavigate, useParams } from "react-router-dom";
 import { fetchEvents } from "../../apis";
 import { EventType } from "../../types";
+import { EventNearHereList, StyledEventNearHere } from "./styles/eventNearHereStyle";
 
 function EventNearHere({ biasesId, district }: Partial<EventType>) {
 	const { id } = useParams();
+	const navigate = useNavigate();
 
 	const { data: nearEvent } = useQuery(["event", id], fetchEvents, {
 		enabled: !!id,
@@ -29,15 +29,19 @@ function EventNearHere({ biasesId, district }: Partial<EventType>) {
 			<h4>가까운 연관 이벤트</h4>
 			<ul>
 				{nearEvent &&
-					nearEvent.map((event) => (
-						<EventNearHereList
-							id={event.id}
-							key={event.id}
-							images={event.images}
-							place={event.place}
-							organizer={event.organizer}
-						/>
-					))}
+					nearEvent.map((event) => {
+						const { eventId, images, place, organizer } = event;
+						const previewUrl = (images && images[0]) || "";
+						return (
+							<EventNearHereList key={eventId} onClick={() => navigate(`/detail/${id}`)}>
+								{previewUrl && <img alt={previewUrl} src={previewUrl} />}
+								<div>
+									<h6>{place}</h6>
+									<p>{organizer}</p>
+								</div>
+							</EventNearHereList>
+						);
+					})}
 			</ul>
 		</StyledEventNearHere>
 	);
