+++ src/components/detail/index.ts
@@ -1,0 +1,8 @@
+import EventMain from "./EventMain";
+import EventNearHere from "./EventNearHere";
+import GoodsInfo from "./GoodsInfo";
+import GoodsListItem from "./GoodsListItem";
+import TwitterInfo from "./TwitterInfo";
+import Location from "./Location";
+
+export { EventMain, EventNearHere, GoodsInfo, GoodsListItem, TwitterInfo, Location };
