+++ src/components/detail/styles/detailStyle.ts
@@ -1,13 +1,13 @@
 import styled from "styled-components";
 
 const StyledDetail = styled.div`
-  @media ${({ theme }) => theme.device.desktop} {
-    display: flex;
+	@media ${({ theme }) => theme.device.desktop} {
+		display: flex;
 
-    > div {
-      width: 50%;
-    }
-  }
+		> div {
+			width: 50%;
+		}
+	}
 `;
 
 export { StyledDetail };
