+++ src/components/detail/styles/goodsInfoStyle.ts
