+++ src/components/detail/styles/locationStyle.ts
