+++ src/components/detail/styles/twitterInfoStyle.ts
