+++ src/components/main/styles/districtSelectorStyle.ts
