--- src/styles/index.ts
@@ -1,6 +1,0 @@
-import { StyledEventMain } from "./eventMainStyle";
-import { StyledTwitterInfo } from "./twitterInfoStyle";
-import { StyledGoodsInfo, StyledGoodsList, StyledGoodsListItem, Chip } from "./goodsInfoStyle";
-
-export { StyledEventMain, StyledTwitterInfo, StyledGoodsInfo, StyledGoodsList, StyledGoodsListItem, Chip };
-export default {};
